moved streams to 'aeon_analysis'
