+++ aeon/aeon_analysis/exp02/__init__.py
@@ -1,0 +1,10 @@
+from importlib_metadata import PackageNotFoundError, version
+
+try:
+    # Change here if project is renamed and does not equal the package name
+    dist_name = "aeon"
+    __version__ = version(dist_name)
+except PackageNotFoundError:
+    __version__ = "unknown"
+finally:
+    del version, PackageNotFoundError
